moved components with css to own folder
