+++ src/components/UnpublishedListing/UnpublishedListing.js
@@ -5,7 +5,7 @@ import { Link } from 'react-router';
 import moment from 'moment';
 import { Card, CardTitle, CardText, CardActions } from 'react-toolbox/lib/card';
 import Button from 'react-toolbox/lib/button';
-import { status, statusDescriptions } from '../constants/publishModes';
+import { status, statusDescriptions } from '../../constants/publishModes';
 import styles from './UnpublishedListing.css';
 
 class UnpublishedListing extends React.Component {
